+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixed
 
+- \#179 \[Dropbox\] \[Google Drive\] Open from Cloud Drive only list partial content
 - \#196 HiDrive: Saving to root folder throws exception